+++ CHANGELOG.md
@@ -8,8 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
+- Resident
+    - Optional email field for contract resident
 - Validation constraint for currency including 0 €
-- Generic `TextInputField` component
+- Generic `TextInputField` component    
 
 ### Changed